booking_clone/requirements.txt
asgiref==3.11.1
attrs==25.4.0
Django==6.0.2
django-debug-toolbar==6.2.0
django-cors-headers==4.3.1
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.1.1
PyJWT==2.11.0
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.2.0
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
google-genai==0.5.0
